cleanup and adding postgres setup

+++ internal/initialize/initialize.go
@@ -90,11 +90,17 @@ write_files:
       systemctl start mariadb
 
       # create the database and user
-      export USER_PASS=$(openssl rand -base64 16)
-      echo "$USER_PASS" > /home/ubuntu/.db_password
-      export MYSQL_USER=$(cat /home/ubuntu/.db_password)
+      if [ -f /home/ubuntu/.db_password ]; then
+          # read the file contents for the password
+          export DB_USER_PASS=$(cat /home/ubuntu/.db_password)
+      else
+          # create a random password and store it
+          export USER_PASS=$(openssl rand -base64 16)
+          echo "$USER_PASS" > /home/ubuntu/.db_password
+          export DB_USER_PASS=$(cat /home/ubuntu/.db_password)
+      fi
 
-      sudo sed -i 's|CHANGEME|'$MYSQL_USER'|g' /opt/nitro/mariadb/init.sql
+      sudo sed -i 's|CHANGEME|'$DB_USER_PASS'|g' /opt/nitro/mariadb/init.sql
 
       sudo mysql -u root < /opt/nitro/mariadb/init.sql
     permissions: '770'
@@ -102,10 +108,37 @@ write_files:
     content: |
       #!/bin/bash
       apt install -y mariadb-server
+  - path: /opt/nitro/postgres/setup.sh
+    content: |
+      # allow remote access to postgres
+      sed -i 's|#listen_addresses = 'localhost'|listen_addresses = '*'|g' /etc/postgresql/10/main/postgresql.conf
+      sed -i 's|127.0.0.1/32|0.0.0.0/0|g' /etc/postgresql/10/main/pg_hba.conf
+      sudo service postgresql restart
+
+      # create the database and user
+      if [ -f /home/ubuntu/.db_password ]; then
+          # read the file contents for the password
+          export DB_USER_PASS=$(cat /home/ubuntu/.db_password)
+      else
+          # create a random password and store it
+          export USER_PASS=$(openssl rand -base64 16)
+          echo "$USER_PASS" > /home/ubuntu/.db_password
+          export DB_USER_PASS=$(cat /home/ubuntu/.db_password)
+      fi
+
+      sudo sed -i 's|CHANGEME|'$DB_USER_PASS'|g' /opt/nitro/postgres/init.sql
+
+      sudo -u postgres psql -U ubuntu < /opt/nitro/postgres/init.sql
+    permissions: '770'
   - path: /opt/nitro/postgres/install.sh
     content: |
       #!/bin/bash
       apt install -y postgresql postgresql-contrib
+  - path: /opt/nitro/postgres/init.sql
+    content: |
+      CREATE DATABASE craftcms;
+      CREATE ROLE craftcms WITH LOGIN PASSWORD 'CHANGEME';
+      GRANT SELECT, INSERT, UPDATE, CREATE, DELETE, REFERENCES, CONNECT ON craftcms TO craftcms; 
   # create nginx install scripts
   - path: /opt/nitro/nginx/install.sh
     content: |
